refactor `E2ETestFixture`
