+++ boudicca.base/publisher-event-html/src/test/kotlin/base/boudicca/publisher/event/html/fixture/E2ETestFixture.kt
@@ -1,4 +1,4 @@
-package base.boudicca.publisher.event.html
+package base.boudicca.publisher.event.html.fixture
 
 import com.microsoft.playwright.*
 import org.junit.jupiter.api.*
